Documents/My Projects/90_north/90_North_Django_/90_North_Django/Django/requirements.txt
absl-py==2.1.0
altair==5.2.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
automata-lib==8.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cached_method==0.1.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpickle==3.0.0
cloup==0.13.1
colorama==0.4.6
colormath==3.0.0
colour==0.1.5
comm==0.1.3
constantly==23.10.4
contourpy==1.1.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
djangorestframework==3.15.2
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
facenet-pytorch==2.6.0
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.43.1
forbiddenfruit==0.1.4
fqdn==1.5.1
frozendict==2.4.0
fsspec==2024.9.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
glcontext==2.5.0
google==3.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
idna==3.4
imageio==2.34.1
incremental==22.10.0
ipykernel==6.23.1
ipython==8.14.0
ipywidgets==8.1.5
isoduration==20.11.0
isosurfaces==0.1.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
Keras-Preprocessing==1.1.2
keras-utils==1.0.13
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.42.0
lxml==5.1.1
manim==0.17.3
ManimPango==0.4.4
mapbox-earcut==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
moderngl==5.8.2
moderngl-window==2.4.4
mpmath==1.3.0
mtcnn==0.1.1
multipledispatch==1.0.0
mysql-connector-python==8.2.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.1
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.1
pandas==2.1.2
pandocfilters==1.5.1
paramiko==3.4.0
parsel==1.9.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.5.1
prometheus_client==0.20.0
prompt-toolkit==3.0.38
Protego==0.3.0
protobuf==4.21.12
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.21
pydeck==0.8.1b0
PyDispatcher==2.0.7
pydot==2.0.0
pydub==0.25.1
pyglet==2.0.9
Pygments==2.15.1
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pyrr==0.10.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==24.0.1
queuelib==1.6.2
referencing==0.33.0
regex==2024.5.10
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.18.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.11.3
Scrapy==2.11.1
screeninfo==0.8.1
seaborn==0.13.2
selenium==4.19.0
Send2Trash==1.8.2
service-identity==24.1.0
shap==0.45.0
six==1.16.0
skia-pathops==0.7.4
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
split-folders==0.5.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
srt==3.5.3
stack-data==0.6.2
streamlit==1.32.0
svgelements==1.9.6
sympy==1.12
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
tf_keras==2.17.0
threadpoolctl==3.4.0
tifffile==2024.5.10
tinycss2==1.2.1
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
visual-automata==1.1.1
w3lib==2.1.2
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
whisper==1.1.10
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.2
